+++ packages/mason/lib/src/bricks_json.dart
@@ -257,7 +257,7 @@ class BricksJson {
     if (directoryExists) await directory.delete(recursive: true);
 
     await directory.parent.create(recursive: true);
-    await tempDirectory.rename(directory.path);
+    await _copyPath(tempDirectory.path, directory.path);
 
     final yaml = _getBrickYaml(directory);
     final name = brick.name ?? yaml.name;
@@ -467,5 +467,17 @@ void _verifyMasonVersionConstraint(BrickYaml brickYaml) {
       brickName: brickYaml.name,
       constraint: brickYaml.environment.mason,
     );
+  }
+}
+
+Future<void> _copyPath(String from, String to) async {
+  await Directory(to).create(recursive: true);
+  await for (final file in Directory(from).list(recursive: true)) {
+    final copyTo = p.join(to, p.relative(file.path, from: from));
+    if (file is Directory) {
+      await Directory(copyTo).create(recursive: true);
+    } else if (file is File) {
+      await File(file.path).copy(copyTo);
+    }
   }
 }