+++ packages/mason_cli/lib/src/commands/new.dart
@@ -48,15 +48,6 @@ class NewCommand extends MasonCommand {
     );
     final createHooks = results['hooks'] == true;
     final directory = Directory(outputDir);
-    final brickYaml = File(p.join(directory.path, name, BrickYaml.file));
-
-    if (brickYaml.existsSync()) {
-      logger.err(
-        'Existing brick: $name at ${canonicalize(brickYaml.parent.path)}',
-      );
-      return ExitCode.usage.code;
-    }
-
     final target = DirectoryGeneratorTarget(directory);
     const vars = <String, dynamic>{'name': '{{name}}'};
     final generator = _BrickGenerator(
@@ -64,19 +55,15 @@ class NewCommand extends MasonCommand {
       description,
       createHooks: createHooks,
     );
-    final newProgress = logger.progress('Creating new brick: $name.');
+    final progress = logger.progress('Creating new brick: $name.');
 
     try {
       await generator.generate(target, vars: vars, logger: logger);
-      newProgress.complete('Created new brick: $name');
-      logger
-        ..info(
-          '''${lightGreen.wrap('✓')} Generated ${generator.files.length} file(s):''',
-        )
-        ..flush((message) => logger.info(darkGray.wrap(message)));
+      progress.complete('Generated ${generator.files.length} file(s).');
+      logger.flush((message) => logger.info(darkGray.wrap(message)));
       return ExitCode.success.code;
     } catch (_) {
-      newProgress.fail();
+      progress.fail();
       rethrow;
     }
   }