+++ packages/mason_cli/lib/src/commands/remove.dart
@@ -42,7 +42,7 @@ class RemoveCommand extends MasonCommand {
     final masonLockJson = isGlobal ? globalMasonLockJson : localMasonLockJson;
 
     final masonYamlFile = isGlobal ? globalMasonYamlFile : localMasonYamlFile;
-    final removeProgress = logger.progress('Removing $brickName');
+    final progress = logger.progress('Removing $brickName');
     try {
       bricksJson.remove(Brick(name: brickName, location: brickLocation));
       final bricks = Map.of(masonYaml.bricks)
@@ -60,9 +60,9 @@ class RemoveCommand extends MasonCommand {
         );
       }
 
-      removeProgress.complete('Removed $brickName');
+      progress.complete('Removed $brickName');
     } catch (_) {
-      removeProgress.fail();
+      progress.fail();
       rethrow;
     }