+++ packages/mason_cli/lib/src/install_brick.dart
@@ -6,44 +6,6 @@ import 'package:path/path.dart' as p;
 
 /// Mixin on [MasonCommand] which adds support for installing bricks.
 mixin InstallBrickMixin on MasonCommand {
-  /// Installs a specific brick and returns the [CachedBrick] reference.
-  Future<CachedBrick> addBrick(Brick brick, {bool global = false}) async {
-    final bricksJson = global ? globalBricksJson : localBricksJson;
-    if (bricksJson == null) {
-      usageException(
-        'Mason has not been initialized.\nDid you forget to run mason init?',
-      );
-    }
-
-    final masonLockJsonFile =
-        global ? globalMasonLockJsonFile : localMasonLockJsonFile;
-    final masonLockJson = global ? globalMasonLockJson : localMasonLockJson;
-    final installProgress = logger.progress('Installing ${brick.name}');
-    try {
-      final location = resolveBrickLocation(
-        location: brick.location,
-        lockedLocation: masonLockJson.bricks[brick.name],
-      );
-      final cachedBrick = await bricksJson.add(
-        Brick(name: brick.name, location: location),
-      );
-      await bricksJson.flush();
-      await masonLockJsonFile.writeAsString(
-        json.encode(
-          MasonLockJson(
-            bricks: {
-              ...masonLockJson.bricks,
-              brick.name!: cachedBrick.brick.location
-            },
-          ),
-        ),
-      );
-      return cachedBrick;
-    } finally {
-      installProgress.complete();
-    }
-  }
-
   /// Installs all bricks either locally or globally depending on [global].
   /// If [upgrade] is true, bricks are upgraded to the latest version
   /// and the lock file is regenerated.
@@ -52,8 +14,6 @@ mixin InstallBrickMixin on MasonCommand {
     if (bricksJson == null) throw const MasonYamlNotFoundException();
     final lockJson = global ? globalMasonLockJson : localMasonLockJson;
     final resolvedBricks = <String, BrickLocation>{};
-    final message = upgrade ? 'Upgrading bricks' : 'Getting bricks';
-    final getBricksProgress = logger.progress(message);
 
     Future<CachedBrick> _resolveBrickEntry(
       MapEntry<String, BrickLocation> entry,
@@ -71,14 +31,12 @@ mixin InstallBrickMixin on MasonCommand {
                 ),
               )
             : location;
-        getBricksProgress.update('Installing ${entry.key}');
         final cachedBrick = await bricksJson.add(
           Brick(name: entry.key, location: normalizedLocation),
         );
         resolvedBricks.addAll(
           <String, BrickLocation>{entry.key: cachedBrick.brick.location},
         );
-        getBricksProgress.update('Building ${entry.key}');
         final generator = await MasonGenerator.fromBrick(
           Brick.path(cachedBrick.path),
         );
@@ -121,7 +79,6 @@ mixin InstallBrickMixin on MasonCommand {
         ]);
       }
     } finally {
-      getBricksProgress.complete(message);
       await bricksJson.flush();
       final masonLockJsonFile =
           global ? globalMasonLockJsonFile : localMasonLockJsonFile;