+++ packages/mason_cli/test/commands/publish_test.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:args/args.dart';
+import 'package:args/command_runner.dart';
 import 'package:mason/mason.dart' hide packageVersion;
 import 'package:mason_api/mason_api.dart';
 import 'package:mason_cli/src/commands/commands.dart';
@@ -10,15 +11,15 @@ import 'package:test/test.dart';
 
 import '../helpers/helpers.dart';
 
-class MockLogger extends Mock implements Logger {}
+class _MockLogger extends Mock implements Logger {}
 
-class MockMasonApi extends Mock implements MasonApi {}
+class _MockMasonApi extends Mock implements MasonApi {}
 
-class MockUser extends Mock implements User {}
+class _MockUser extends Mock implements User {}
 
-class MockArgResults extends Mock implements ArgResults {}
+class _MockArgResults extends Mock implements ArgResults {}
 
-class MockProgress extends Mock implements Progress {}
+class _MockProgress extends Mock implements Progress {}
 
 class FakeUri extends Fake implements Uri {}
 
@@ -38,14 +39,14 @@ void main() {
     late PublishCommand publishCommand;
 
     setUp(() async {
-      logger = MockLogger();
-      masonApi = MockMasonApi();
-      argResults = MockArgResults();
+      logger = _MockLogger();
+      masonApi = _MockMasonApi();
+      argResults = _MockArgResults();
       publishCommand = PublishCommand(
         logger: logger,
         masonApiBuilder: ({Uri? hostedUri}) => masonApi,
       )..testArgResults = argResults;
-      when(() => logger.progress(any())).thenReturn(MockProgress());
+      when(() => logger.progress(any())).thenReturn(_MockProgress());
       setUpTestingEnvironment(cwd, suffix: '.publish');
     });
 
@@ -55,6 +56,23 @@ void main() {
 
     test('can be instantiated without any parameters', () {
       expect(PublishCommand.new, returnsNormally);
+    });
+
+    test(
+        'throws usageException when '
+        'called with both --force and --dry-run', () async {
+      final runner = CommandRunner<int>('mason', '')
+        ..addCommand(publishCommand..testArgResults = null);
+      await expectLater(
+        () => runner.run(['publish', '--dry-run', '--force']),
+        throwsA(
+          isA<UsageException>().having(
+            (e) => e.message,
+            'message',
+            'Cannot use both --force and --dry-run.',
+          ),
+        ),
+      );
     });
 
     test('exits with code 70 when brick could not be found', () async {
@@ -122,7 +140,7 @@ Please change or remove the "publish_to" field in the brick.yaml before publishi
     });
 
     test('exits with code 70 when email is not verified', () async {
-      final user = MockUser();
+      final user = _MockUser();
       when(() => user.emailVerified).thenReturn(false);
       when(() => masonApi.currentUser).thenReturn(user);
       when(() => argResults['directory'] as String).thenReturn(brickPath);
@@ -135,7 +153,7 @@ Please change or remove the "publish_to" field in the brick.yaml before publishi
     });
 
     test('exits with code 70 when bundle is too large', () async {
-      final user = MockUser();
+      final user = _MockUser();
       when(() => user.emailVerified).thenReturn(true);
       when(() => masonApi.currentUser).thenReturn(user);
       when(() => argResults['directory'] as String).thenReturn(brickPath);
@@ -157,11 +175,45 @@ Please change or remove the "publish_to" field in the brick.yaml before publishi
       verify(() => masonApi.close()).called(1);
     });
 
+    test('exits with code 0 without publishing when using --dry-run', () async {
+      final user = _MockUser();
+      final progressLogs = <String>[];
+      when(() => user.emailVerified).thenReturn(true);
+      when(() => masonApi.currentUser).thenReturn(user);
+      when(
+        () => masonApi.publish(bundle: any(named: 'bundle')),
+      ).thenAnswer((_) async {});
+      final progress = _MockProgress();
+      when(() => progress.complete(any())).thenAnswer((invocation) {
+        final update = invocation.positionalArguments[0] as String?;
+        if (update != null) progressLogs.add(update);
+      });
+      when(() => logger.progress(any())).thenReturn(progress);
+      when(() => logger.confirm(any())).thenReturn(true);
+      when(() => argResults['directory'] as String).thenReturn(brickPath);
+      when(() => argResults['dry-run'] as bool).thenReturn(true);
+      final result = await publishCommand.run();
+      expect(result, equals(ExitCode.success.code));
+      expect(progressLogs, equals(['Bundled greeting']));
+      verify(() => logger.info('No issues detected.')).called(1);
+      verify(
+        () => logger.info('The server may enforce additional checks.'),
+      ).called(1);
+      verifyNever(() => logger.progress('Publishing greeting 0.1.0+1'));
+      verifyNever(() {
+        logger.success(
+          '''\nPublished greeting 0.1.0+1 to ${BricksJson.hostedUri}.''',
+        );
+      });
+      verifyNever(() => masonApi.publish(bundle: any(named: 'bundle')));
+      verify(() => masonApi.close()).called(1);
+    });
+
     test('exits with code 70 when publish is aborted', () async {
       final policyLink = styleUnderlined.wrap(
         link(uri: Uri.parse('https://brickhub.dev/policy')),
       );
-      final user = MockUser();
+      final user = _MockUser();
       when(() => user.emailVerified).thenReturn(true);
       when(() => masonApi.currentUser).thenReturn(user);
       when(() => logger.confirm(any())).thenReturn(false);
@@ -191,7 +243,7 @@ Please change or remove the "publish_to" field in the brick.yaml before publishi
     });
 
     test('exits with code 70 when publish fails', () async {
-      final user = MockUser();
+      final user = _MockUser();
       const message = 'oops';
       const exception = MasonApiPublishFailure(message: message);
       when(() => user.emailVerified).thenReturn(true);
@@ -211,7 +263,7 @@ Please change or remove the "publish_to" field in the brick.yaml before publishi
 
     test('exits with code 70 when publish fails (generic)', () async {
       final exception = Exception('oops');
-      final user = MockUser();
+      final user = _MockUser();
       when(() => user.emailVerified).thenReturn(true);
       when(() => masonApi.currentUser).thenReturn(user);
       when(
@@ -226,14 +278,14 @@ Please change or remove the "publish_to" field in the brick.yaml before publishi
     });
 
     test('exits with code 0 when publish succeeds', () async {
-      final user = MockUser();
+      final user = _MockUser();
       final progressLogs = <String>[];
       when(() => user.emailVerified).thenReturn(true);
       when(() => masonApi.currentUser).thenReturn(user);
       when(
         () => masonApi.publish(bundle: any(named: 'bundle')),
       ).thenAnswer((_) async {});
-      final progress = MockProgress();
+      final progress = _MockProgress();
       when(() => progress.complete(any())).thenAnswer((invocation) {
         final update = invocation.positionalArguments[0] as String?;
         if (update != null) progressLogs.add(update);
@@ -257,13 +309,46 @@ Please change or remove the "publish_to" field in the brick.yaml before publishi
       verify(() => masonApi.close()).called(1);
     });
 
+    test('exits with code 0 when publish succeeds with --force', () async {
+      final user = _MockUser();
+      final progressLogs = <String>[];
+      when(() => user.emailVerified).thenReturn(true);
+      when(() => masonApi.currentUser).thenReturn(user);
+      when(
+        () => masonApi.publish(bundle: any(named: 'bundle')),
+      ).thenAnswer((_) async {});
+      final progress = _MockProgress();
+      when(() => progress.complete(any())).thenAnswer((invocation) {
+        final update = invocation.positionalArguments[0] as String?;
+        if (update != null) progressLogs.add(update);
+      });
+      when(() => logger.progress(any())).thenReturn(progress);
+      when(() => argResults['directory'] as String).thenReturn(brickPath);
+      when(() => argResults['force'] as bool).thenReturn(true);
+      final result = await publishCommand.run();
+      expect(result, equals(ExitCode.success.code));
+      expect(
+        progressLogs,
+        equals(['Bundled greeting', 'Published greeting 0.1.0+1']),
+      );
+      verify(() => logger.progress('Publishing greeting 0.1.0+1')).called(1);
+      verify(() {
+        logger.success(
+          '''\nPublished greeting 0.1.0+1 to ${BricksJson.hostedUri}.''',
+        );
+      }).called(1);
+      verifyNever(() => logger.confirm(any()));
+      verify(() => masonApi.publish(bundle: any(named: 'bundle'))).called(1);
+      verify(() => masonApi.close()).called(1);
+    });
+
     test(
       'exits with code 0 when publish succeeds '
       'with a custom registry (publish_to)',
       () async {
         final brickPath = p.join('..', '..', 'bricks', 'custom_registry');
         final customHostedUri = Uri.parse('https://custom.brickhub.dev');
-        final user = MockUser();
+        final user = _MockUser();
         final progressLogs = <String>[];
         when(() => user.emailVerified).thenReturn(true);
 
@@ -280,7 +365,7 @@ Please change or remove the "publish_to" field in the brick.yaml before publishi
         when(
           () => masonApi.publish(bundle: any(named: 'bundle')),
         ).thenAnswer((_) async {});
-        final progress = MockProgress();
+        final progress = _MockProgress();
         when(() => progress.complete(any())).thenAnswer((invocation) {
           final update = invocation.positionalArguments[0] as String?;
           if (update != null) progressLogs.add(update);