+++ packages/mason/CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.1.0-dev.48
+
+- fix: git brick install across file systems
+- refactor: use `Isolate.run`
+
+# 0.1.0-dev.47
+
+- fix: loop detection in `runSubstitution`
+- test: use private mocks
+
 # 0.1.0-dev.46
 
 - fix: `MasonBundle` use `fieldRename: FieldRename.snake`